requirements.txt
# FastAPI Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
stripe==7.10.0
openai==1.7.2
qrcode[pil]==7.4.2
redis==5.0.1
httpx==0.26.0
